Python graphing script now sets number of particles based on the size of bin files.  Fixed a couple of very minor file, folder name bugs.  Replaced some cout \n's with std::endl's. Pre-mild restructure (GPU functions to Particle class).

+++ src/FileIO/fileIO.cpp
@@ -18,7 +18,7 @@ namespace fileIO
 		}
 		
 		binFile.seekg(0, binFile.end);
-		int length{ binFile.tellg() };
+		int length{ static_cast<int>(binFile.tellg()) };
 		binFile.seekg(0, binFile.beg);
 
 		if (length < numOfDblsToRead * 8)
